pyproject.toml
[tool.poetry]
name = "pyogmios"
version = "0.7.3"
description = "Python client for Ogmios"
authors = ["Hareem Adderley <hadderley@kingpinapps.com>"]
readme = "README.md"
packages = [{include = "pyogmios_client"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta"
]

[project.urls]
"Homepage" = "https://github.com/KINGH242/pyogmios"
"Bug Tracker" = "https://github.com/KINGH242/pyogmios/issues"

[tool.poetry.dependencies]
python = ">=3.8.10, <3.12"
requests = "^2.31.0"
pydantic = "^2.5.1"
aiohttp = "^3.9.0"
aiodns = "^3.1.1"
websocket-client = "^1.6.4"
promise = "^2.3"
nanoid = "^2.0.0"
pyee = "^11.0.1"
rel = "^0.4.9"


[tool.poetry.group.dev.dependencies]
datamodel-code-generator = "^0.24.2"
black = "^22.6.0"
commitizen = "^2.32.5"
coverage = "^6.4.4"
Faker = "^15.3.4"
flake8 = "^5.0.4"
pylint = "^2.14.5"
pytest = "^7.1.3"
pytest-asyncio = "^0.20.3"
pytest-mock = "^3.8.2"
ruff = "^0.0.261"
polyfactory = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-p no:cacheprovider"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.7.3"
version_files = [
    "pyogmios_client/__init__.py:__version__",
    "pyproject.toml:version"
]
update_changelog_on_bump = true
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]

[tool.ruff]
ignore = [
    "E501",
]

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
